Changes in nextcloud - adding not static link but dynamic and removing the whitespaces.

+++ scripts/chce_nextcloud.sh
@@ -1,16 +1,15 @@
 #!/bin/bash
 # NEXTCLOUD installation script
-# Autor: Mariusz 'maniek205' Kowalski
+# Autors: Mariusz 'maniek205' Kowalski, Marcin Wozniak
 
 USERNAME=admin
-PASSWORD=$(head -c 100 /dev/urandom | tr -dc A-Za-z0-9 | head -c13) 
+PASSWORD=$(head -c 100 /dev/urandom | tr -dc A-Za-z0-9 | head -c13)
 
 DB_USER=root
-DB_PASS=$(head -c 100 /dev/urandom | tr -dc A-Za-z0-9 | head -c13) 
+DB_PASS=$(head -c 100 /dev/urandom | tr -dc A-Za-z0-9 | head -c13)
 
 #Set Timezone to prevent installation interruption
-ln -snf /usr/share/zoneinfo/Poland /etc/localtime && echo "Etc/UTC" > /etc/timezone
-
+[[ ! -f /etc/localtime ]] && ln -snf /usr/share/zoneinfo/Poland /etc/localtime && echo "Etc/UTC" > /etc/timezone
 
 #Installing prerequisites https://docs.nextcloud.com/server/latest/admin_manual/installation/example_ubuntu.html
 apt update
@@ -20,15 +19,21 @@ apt install -y php7.4-gmp php7.4-bcmath php-imagick php7.4-xml php7.4-zip
 
 #Configuring mariaDB
 /etc/init.d/mysql start
-mysql -u$DB_USER -p$DB_PASS -e "CREATE USER '$USERNAME'@'localhost' IDENTIFIED BY '$PASSWORD'; 
-CREATE DATABASE IF NOT EXISTS nextcloud CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci; 
-GRANT ALL PRIVILEGES ON nextcloud.* TO '$USERNAME'@'localhost'; 
+mysql -u"$DB_USER" -p"$DB_PASS" -e "CREATE USER '$USERNAME'@'localhost' IDENTIFIED BY '$PASSWORD';
+CREATE DATABASE IF NOT EXISTS nextcloud CHARACTER SET utf8mb4 COLLATE utf8mb4_general_ci;
+GRANT ALL PRIVILEGES ON nextcloud.* TO '$USERNAME'@'localhost';
 FLUSH PRIVILEGES;"
 
-#Downloading nextcloud zip file
-apt install -y wget tar
-wget https://download.nextcloud.com/server/releases/nextcloud-22.2.0.tar.bz2
-tar -xf nextcloud-22.2.0.tar.bz2
+#Downloading nextcloud tar.bz2 file
+apt install -y wget tar curl
+
+nextcloud_link=$(curl https://nextcloud.com/install/\#instructions-server \
+	| grep -Eo 'https://.+\/releases\/.+\.tar\.bz2"' | sed 's/"//g')
+nextcloud_tmp="/tmp/nextcloud.tar.bz2"
+
+wget "$nextcloud_link" -O "$nextcloud_tmp"
+tar -xf "$nextcloud_tmp"
+
 #Copy nextcloud to apache folder
 rm /var/www/html/index.html
 cp -r nextcloud/ /var/www/html/
@@ -58,6 +63,8 @@ a2enmod mime
 a2enmod setenvif
 service apache2 reload
 service apache2 restart
+
+rm -v "$nextcloud_tmp"
 
 echo "USERNAME=$USERNAME
 PASSWORD=$PASSWORD
